+++ api/problems.go
@@ -4,19 +4,24 @@ import (
 	"encoding/json"
 	"log"
 	"net/http"
+	"racegex/database"
+	"racegex/models"
 
 	"github.com/gorilla/mux"
 )
 
 // GetProblems gets the list of all the problems
 func GetProblems(w http.ResponseWriter, r *http.Request) {
-	var data []Problem
+	problems := []models.Problem{}
 
-	data = append(data, Problem{ID: "1", Title: "emails"})
-	data = append(data, Problem{ID: "2", Title: "websites"})
+	res := database.DBConn.Find(&problems)
 
-	if err := json.NewEncoder(w).Encode(data); err != nil {
-		log.Printf("Error parsing %-v", data)
+	if res.Error != nil {
+		log.Printf("Couldn't get problems")
+	}
+
+	if err := json.NewEncoder(w).Encode(res.Value); err != nil {
+		log.Printf("Error parsing %-v", problems)
 	}
 }
 
@@ -25,25 +30,15 @@ func GetProblem(w http.ResponseWriter, r *http.Request) {
 	// Get the url params of the route
 	params := mux.Vars(r)
 
-	// Initialize a mock array of items to look for the id
-	var data []Problem
+	problem := models.Problem{}
 
-	data = append(data, Problem{ID: "1", Title: "emails"})
-	data = append(data, Problem{ID: "2", Title: "websites"})
+	res := database.DBConn.First(&problem, params["id"])
 
-	for _, problem := range data {
-		if problem.ID == params["id"] {
-			if err := json.NewEncoder(w).Encode(problem); err != nil {
-				log.Printf("Error parsing %-v", problem)
-			}
-			return
-		}
+	if res.Error != nil {
+		log.Printf("Couldn't get problem %-v", params["id"])
 	}
-	json.NewEncoder(w).Encode(Problem{})
-}
 
-// Problem : Contains a problem to challenge the user with
-type Problem struct {
-	ID    string `json:"id"`
-	Title string `json:"title"`
+	if err := json.NewEncoder(w).Encode(res.Value); err != nil {
+		log.Printf("Error parsing %-v", problem)
+	}
 }