+++ CHANGELOG.md
@@ -32,5 +32,6 @@ Please check https://keepachangelog.com/en/1.0.0/#how
 - Create an initial directory structure that handles routes in `main.go`, but
   delegates all implementations of handlers to `api` and `socket` subpackages.
 - Allow CORS.
+- Connect with a PostgreSQL database and seed it with automigrated models.
 
 [`unreleased`]: https://github.com/daque-dev/sv-racegex/tree/develop